% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercor.R
\name{intercor}
\alias{intercor}
\title{Computes a matrix of Spearman's rho or Pearson's r rank correlation coefficients for all possible pairs of columns of a matrix, filter based on the threshold of P and r cutoff, and output the matrix.}
\usage{
intercor(x, y, method, pcutoff, rcutoff)
}
\arguments{
\item{x}{a numeric matrix}

\item{y}{a numeric matrix}

\item{method}{specifies the type of correlations to compute}

\item{pcutoff}{Specifies a threshold for the p value}

\item{rcutoff}{Specifies a threshold for the r value}
}
\description{
Computes a matrix of Spearman's rho or Pearson's r rank correlation coefficients for all possible pairs of columns of a matrix, filter based on the threshold of P and r cutoff, and output the matrix.
}
